+++ CHANGELOG.md
@@ -4,10 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+* Add `utils.Key` (returned by redesigned `utils.getch`)
 ### Changed
 * Use links of the [virtual campus](https://campusvirtual.uva.es/my) instead of [google drive](https://drive.google.com/drive).
 * Renamed `IconType.unkown` to `IconType.unknown` (**typo**).
-* Redesigned `utils.getch`.
+* Redesigned `utils.getch`. Now returns `utils.Key`.
 
 ### Removed
 * Removed `IconType.not_id`.