+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+* Add `utils.Key` (returned by redesigned `utils.getch`)
+### Changed
+* Use links of the [virtual campus](https://campusvirtual.uva.es/my) instead of [google drive](https://drive.google.com/drive).
+* Renamed `IconType.unkown` to `IconType.unknown` (**typo**).
+* Redesigned `utils.getch`. Now returns `utils.Key`.
+
+### Removed
+* Removed `IconType.not_id`.
+
 
 ## [3.0.2] - 2020-03-13
 ### Added