+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [3.1.0] - 2020-03-23
 ### Added
 * Add `utils.Key` (returned by redesigned `utils.getch`).
 * Add support for `Chat` links.
@@ -68,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix notifier error.
 
 
-[unreleased]: https://github.com/sralloza/vcm/compare/v3.0.2...HEAD
+[unreleased]: https://github.com/sralloza/vcm/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/sralloza/vcm/compare/v3.0.2...v3.1.0
 [3.0.2]: https://github.com/sralloza/vcm/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/sralloza/vcm/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/sralloza/vcm/compare/v2.1.1...v3.0.0